test(ccr-core): split npu_smoke into two test binaries to avoid OnceCell cross-pollution
Both tests touch the OV_EMBEDDER OnceCell. When run together in one process the fall-back test (which sets OPENVINO_LIB_PATH=/dev/null) caches None first; the actually-uses-NPU test then can\047t recover.

Each integration test file is its own crate, so moving the fall-back test into npu_fallback.rs gives them separate processes and isolates OnceCell state.
